Restructure folders to follow 3 project refs
