+++ babel.config.js
@@ -1,3 +1,8 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
 module.exports = (api) => {
     const isTest = api.env("test");
