+++ src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "target": "ES5",
+        "outDir": "../dist"
+    },
+    "include": ["./**/*.ts"]
+}
