+++ tests/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "target": "ESNext",
+        "outDir": "../dist/__tests__",
+        "jsx": "react",
+        "jsxFactory": "BroTest.createElementString"
+    },
+    "include": ["./**/*.ts", "./**/*.tsx"],
+    "references": [{ "path": "../src" }]
+}
