+++ tsconfig.base.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+        "declaration": true,
+        "composite": true,
+        "noResolve": false,
+        "module": "ES2015",
+        "moduleResolution": "Node",
+        "noImplicitAny": true,
+        "noImplicitReturns": true,
+        "noImplicitThis": true,
+        "noUnusedLocals": true,
+        "skipLibCheck": true,
+        "strictNullChecks": true,
+        "strictPropertyInitialization": true,
+        "forceConsistentCasingInFileNames": true,
+        "lib": ["ES5", "ES2015.Iterable", "DOM", "ESNext"],
+        "esModuleInterop": true
+    },
+    "exclude": ["dist", "node_modules"]
+}
