+++ tsconfig.json
@@ -1,27 +1,8 @@
 {
-    "compilerOptions": {
-        "declaration": true,
-        "noResolve": false,
-        "module": "ES2015",
-        "target": "ES5",
-        "moduleResolution": "Node",
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "noImplicitThis": true,
-        "noUnusedLocals": true,
-        "skipLibCheck": true,
-        "strictNullChecks": true,
-        "strictPropertyInitialization": true,
-        "forceConsistentCasingInFileNames": true,
-        "outDir": "./dist/",
-        "jsx": "react",
-        "jsxFactory": "BroTest.createElementString",
-        "lib": ["es5", "es2015.iterable", "dom", "esnext"],
-        "esModuleInterop": true,
-        "rootDir": "./src"
-    },
-
-    "include": ["./src/**/*"],
-
-    "exclude": ["dist", "node_modules"]
+    "references": [
+        { "path": "src" },
+        { "path": "stories" },
+        { "path": "tests" },
+        { "path": "./tsconfig.eslint.json" }
+    ]
 }
